requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==0.19.0
requests>=2.27.0
tweepy>=4.14.0
apscheduler==3.9.1